create contactList folder and move files there
