+++ frontend/ui/src/index.tsx
@@ -1,6 +1,6 @@
 export { Header } from "./Header";
-export { ContactList } from "./components/ContactList";
 export { StepperDrawer } from "./components/StepperDrawer/StepperDrawer";
 export { Slide } from "./components/Slide/Slide";
 export { PreviewLayoutContainer } from "./components/PreviewLayout/PreviewLayoutContainer";
+export { ContactList } from "./components/ContactList/ContactList";
 export { GenericStepper } from "./components/Stepper/GenericStepper";
